Updated required libdwarf version, clarified stopThread snippet description.
Since the last release Dyninst has come to rely on a newer version
of libdwarf (the 20080615 release) and this was not yet specified
in the Dyninst manual (which requests the 20060327 release).  I also
clarified the documentation to state that the stopThread snippet type
is only implemented for the x86-32 bit platform.